Update README.md
Updated instructions for building the JUCE project and transitioning from Faust to Juce to use global paths for modules rather than creating local copies. Added links to most current version of files within the "Building the Code" section. Highlighted sections of code within the "How it works" section and added the specific objects from the oscillators library that are used